Update app.py
This not all of the code this what was not saved just missing the htttp code but no longer utilizing this and much more comfort doing it from vscode then this. I am the one who figured this out finally yesterday deleted quite a few times yesterday if I used the restroom or anything. Great thing is I found this last night and made a couple changes I don't believe this is an accurate form to do this but at least it's better than nothing. All screenshots have been taken and saved since U found this one by luck but this is already done in vscode just diing this ptior to pushing code up since was unaware how this course did it. only have directions of the zip file with bin for privacy that's definitelt;y best option.